\name{coef.elca}
\alias{coef.elca}
\title{
Extract extended Lee-Carter coefficients from an object of class elca
}
\description{
Extract extended Lee-Carter coefficients from an object of class elca
}
\usage{
\method{coef}{elca}(object, ...)
}
\arguments{
  \item{object}{Data object of class elca}
  \item{...}{Other arguments}
}
\author{
Zoltan Butt, Steven Haberman and Han Lin Shang
}
\seealso{
\code{\link{print.elca}}
}
\keyword{method}
