\name{coef.lca}
\alias{coef.lca}
\title{
Extract Lee-Carter coefficients from an object of class lca
}
\description{
Extract Lee-Carter coefficients from an object of class lca
}
\usage{
\method{coef}{lca}(object, ...)
}
\arguments{
  \item{object}{Data object of class lca}
  \item{\dots}{Other arguments}
}
\author{
Zoltan Butt, Steven Haberman and Han Lin Shang
}
\seealso{
\code{\link{coef.elca}}
}
\keyword{method}
