\name{coef.rh}
\alias{coef.rh}
\title{
Extract Lee-Carter coefficients from an object of class rh
}
\description{
Extract Lee-Carter coefficients from an object of class rh.
}
\usage{
\method{coef}{rh}(object, ...)
}
\arguments{
  \item{object}{Data object of class rh}
  \item{...}{Other arguments}
}
\author{
Zoltan Butt, Steven Haberman and Han Lin Shang
}
\seealso{
\code{\link{print.rh}}
}
\keyword{method}
