\name{dd.cmi.pens}
\alias{dd.cmi.pens}
\docType{data}
\title{
Male mortality data of UK pensioners provided by Continuous Mortality Investigation UK
}
\description{
The dataset is made up by age- and time-specific mortality (hazard) rates and population (exposure) for male pensioners in the UK.  Specifically, it covers individual ages between 50 - 108 and calendar years between 1983 - 2003.
}
\usage{data("dd.cmi.pens")}
\format{
Object of class \code{demogdata}
}
\source{
Continuous Mortality Investigation (http://www.actuaries.org.uk/research-and-resources/pages/continuous-mortality-investigation)
}
\seealso{
\code{\link[demography]{demogdata}}
}
\examples{
# print data summary:
dd.cmi.pens
#Mortality data for CMI
#    Series: male
#    Years: 1983 - 2003
#    Ages:  50 - 108 
}
\keyword{datasets}
