\name{fill.rhdata}
\alias{fill.rhdata}
\title{
Interpolate/Smooth an object of class rhdata
}
\description{
Interpolate/Smooth an object of class rhdata
}
\usage{
fill.rhdata(data, method = c("mspline", "interpolate", "perks"), ...)
}
\arguments{
  \item{data}{Data object of class rh}
  \item{method}{Method of interpolation or smoothing}
  \item{\dots}{Other arguments}
}
\value{
An object does not have missing values
}
\author{
Zoltan Butt, Steven Haberman and Han Lin Shang
}
\keyword{method}
