\name{plot.elca}
\alias{plot.elca}
\title{
Plot an object of class elca
}
\description{
Plot an object of class elca
}
\usage{
\method{plot}{elca}(x, ...)
}
\arguments{
  \item{x}{An object of class elca}
  \item{\dots}{Other arguments}
}
\value{
A plot
}
\author{
Zoltan Butt, Steven Haberman and Han Lin Shang
}
\seealso{
\code{\link{print.elca}}
}
\keyword{plot}
