\name{plot.rh}
\alias{plot.rh}
\title{
Plot an object of class rh
}
\description{
Plot an object of class rh
}
\usage{
\method{plot}{rh}(x, ...)
}
\arguments{
  \item{x}{An object of class rh}
  \item{\dots}{Other arguments}
}
\value{
A plot
}
\author{
Zoltan Butt, Steven Haberman and Han Lin Shang
}
\seealso{
\code{\link{print.rh}}
}
\keyword{plot}
