\name{print.elca}
\alias{print.elca}
\title{
Print function for class elca
}
\description{
Print an object of class elca
}
\usage{
\method{print}{elca}(x, ...)
}
\arguments{
  \item{x}{Data object of class elca}
  \item{...}{Other arguments}
}
\author{
Zoltan Butt, Steven Haberman and Han Lin Shang
}
\seealso{
\code{\link{print.rh}}
}
\keyword{method}
