\name{print.rh}
\alias{print.rh}
\title{
Print function for class rh
}
\description{
Print an object of class rhdata
}
\usage{
\method{print}{rh}(x,...)
}
\arguments{
  \item{x}{Data object of class rh}
  \item{...}{Other arguments}
}
\author{
Zoltan Butt, Steven Haberman and Han Lin Shang
}
\seealso{
\code{\link{print.rhdata}}
}
\keyword{method}
