\name{print.rhdata}
\alias{print.rhdata}
\title{
Print function for class rhdata
}
\description{
Print an object of class rhdata
}
\usage{
\method{print}{rhdata}(x, ...)
}
\arguments{
  \item{x}{Data object of class rhdata}
  \item{...}{Other arguments}
}
\author{
Zoltan Butt, Steven Haberman and Han Lin Shang
}
\seealso{
\code{\link{print.rh}}
}
\keyword{method}